// src/proto/dapr/proto/runtime/v1/dapr.ts
export interface RegisterActorTimerRequest {
  actorType: string;
  actorId: string;
  name: string;
  dueTime: string;
  period: string;
  callback: string;
  data: Uint8Array;
  ttl: string;
}

/** UnregisterActorTimerRequest is the message to unregister an actor timer */
export interface UnregisterActorTimerRequest {
  actorType: string;
  actorId: string;
  name: string;
}

/** RegisterActorReminderRequest is the message to register a reminder for an actor of a given type and id. */
export interface RegisterActorReminderRequest {
  actorType: string;
  actorId: string;
  name: string;
  dueTime: string;
  period: string;
  data: Uint8Array;
  ttl: string;
}

/** UnregisterActorReminderRequest is the message to unregister an actor reminder. */
export interface UnregisterActorReminderRequest {
  actorType: string;
  actorId: string;
  name: string;
}

/** RenameActorReminderRequest is the message to rename an actor reminder. */
export interface RenameActorReminderRequest {
  actorType: string;
  actorId: string;
  oldName: string;
  newName: string;
}

/** GetActorStateRequest is the message to get key-value states from specific actor. */
export interface GetActorStateRequest {
  actorType: string;
  actorId: string;
  key: string;
}

/** GetActorStateResponse is the response conveying the actor's state value. */
export interface GetActorStateResponse {
  data: Uint8Array;
  /** The metadata which will be sent to app. */
  metadata: { [key: string]: string };
}

export interface GetActorStateResponse_MetadataEntry {
  key: string;
  value: string;
}

/** ExecuteActorStateTransactionRequest is the message to execute multiple operations on a specified actor. */
export interface ExecuteActorStateTransactionRequest {
  actorType: string;
  actorId: string;
  operations: TransactionalActorStateOperation[];
}

/** TransactionalActorStateOperation is the message to execute a specified operation with a key-value pair. */
export interface TransactionalActorStateOperation {
  operationType: string;
  key: string;
  value:
    | Any
    | undefined;
  /**
   * The metadata used for transactional operations.
   *
   * Common metadata property:
   * - ttlInSeconds : the time to live in seconds for the stored value.
   */
  metadata: { [key: string]: string };
}

export interface TransactionalActorStateOperation_MetadataEntry {
  key: string;
  value: string;
}

/** InvokeActorRequest is the message to call an actor. */
export interface InvokeActorRequest {
  actorType: string;
  actorId: string;
  method: string;
  data: Uint8Array;
  metadata: { [key: string]: string };
}

export interface InvokeActorRequest_MetadataEntry {
  key: string;
  value: string;
}

/** InvokeActorResponse is the method that returns an actor invocation response. */
export interface InvokeActorResponse {
  data: Uint8Array;
}

/** GetMetadataResponse is a message that is returned on GetMetadata rpc call */
export interface GetMetadataResponse {
  id: string;
  activeActorsCount: ActiveActorsCount[];
  registeredComponents: RegisteredComponents[];
  extendedMetadata: { [key: string]: string };
  subscriptions: PubsubSubscription[];
  httpEndpoints: MetadataHTTPEndpoint[];
  appConnectionProperties: AppConnectionProperties | undefined;
  runtimeVersion: string;
  enabledFeatures: string[];
}

export interface GetMetadataResponse_ExtendedMetadataEntry {
  key: string;
  value: string;
}

export interface ActiveActorsCount {
  type: string;
  count: number;
}

export interface RegisteredComponents {
  name: string;
  type: string;
  version: string;
  capabilities: string[];
}

export interface MetadataHTTPEndpoint {
  name: string;
}

export interface AppConnectionProperties {
  port: number;
  protocol: string;
  channelAddress: string;
  maxConcurrency: number;
  health: AppConnectionHealthProperties | undefined;
}

export interface AppConnectionHealthProperties {
  healthCheckPath: string;
  healthProbeInterval: string;
  healthProbeTimeout: string;
  healthThreshold: number;
}

export interface PubsubSubscription {
  pubsubName: string;
  topic: string;
  metadata: { [key: string]: string };
  rules: PubsubSubscriptionRules | undefined;
  deadLetterTopic: string;
}

export interface PubsubSubscription_MetadataEntry {
  key: string;
  value: string;
}

export interface PubsubSubscriptionRules {
  rules: PubsubSubscriptionRule[];
}
// ...
export const GetActorStateRequest = {
  encode(message: GetActorStateRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.actorType !== "") {
      writer.uint32(10).string(message.actorType);
    }
    if (message.actorId !== "") {
      writer.uint32(18).string(message.actorId);
    }
    if (message.key !== "") {
      writer.uint32(26).string(message.key);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): GetActorStateRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetActorStateRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.actorType = reader.string();
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.actorId = reader.string();
          continue;
        case 3:
          if (tag !== 26) {
            break;
          }

          message.key = reader.string();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  // encodeTransform encodes a source of message objects.
  // Transform<GetActorStateRequest, Uint8Array>
  async *encodeTransform(
    source:
      | AsyncIterable<GetActorStateRequest | GetActorStateRequest[]>
      | Iterable<GetActorStateRequest | GetActorStateRequest[]>,
  ): AsyncIterable<Uint8Array> {
    for await (const pkt of source) {
      if (globalThis.Array.isArray(pkt)) {
        for (const p of (pkt as any)) {
          yield* [GetActorStateRequest.encode(p).finish()];
        }
      } else {
        yield* [GetActorStateRequest.encode(pkt as any).finish()];
      }
    }
  },

  // decodeTransform decodes a source of encoded messages.
  // Transform<Uint8Array, GetActorStateRequest>
  async *decodeTransform(
    source: AsyncIterable<Uint8Array | Uint8Array[]> | Iterable<Uint8Array | Uint8Array[]>,
  ): AsyncIterable<GetActorStateRequest> {
    for await (const pkt of source) {
      if (globalThis.Array.isArray(pkt)) {
        for (const p of (pkt as any)) {
          yield* [GetActorStateRequest.decode(p)];
        }
      } else {
        yield* [GetActorStateRequest.decode(pkt as any)];
      }
    }
  },

  fromJSON(object: any): GetActorStateRequest {
    return {
      actorType: isSet(object.actorType) ? globalThis.String(object.actorType) : "",
      actorId: isSet(object.actorId) ? globalThis.String(object.actorId) : "",
      key: isSet(object.key) ? globalThis.String(object.key) : "",
    };
  },

  toJSON(message: GetActorStateRequest): unknown {
    const obj: any = {};
    if (message.actorType !== "") {
      obj.actorType = message.actorType;
    }
    if (message.actorId !== "") {
      obj.actorId = message.actorId;
    }
    if (message.key !== "") {
      obj.key = message.key;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GetActorStateRequest>, I>>(base?: I): GetActorStateRequest {
    return GetActorStateRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GetActorStateRequest>, I>>(object: I): GetActorStateRequest {
    const message = createBaseGetActorStateRequest();
    message.actorType = object.actorType ?? "";
    message.actorId = object.actorId ?? "";
    message.key = object.key ?? "";
    return message;
  },
};
// ...
export const ExecuteActorStateTransactionRequest = {
  encode(message: ExecuteActorStateTransactionRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.actorType !== "") {
      writer.uint32(10).string(message.actorType);
    }
    if (message.actorId !== "") {
      writer.uint32(18).string(message.actorId);
    }
    for (const v of message.operations) {
      TransactionalActorStateOperation.encode(v!, writer.uint32(26).fork()).ldelim();
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): ExecuteActorStateTransactionRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseExecuteActorStateTransactionRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.actorType = reader.string();
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.actorId = reader.string();
          continue;
        case 3:
          if (tag !== 26) {
            break;
          }

          message.operations.push(TransactionalActorStateOperation.decode(reader, reader.uint32()));
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  // encodeTransform encodes a source of message objects.
  // Transform<ExecuteActorStateTransactionRequest, Uint8Array>
  async *encodeTransform(
    source:
      | AsyncIterable<ExecuteActorStateTransactionRequest | ExecuteActorStateTransactionRequest[]>
      | Iterable<ExecuteActorStateTransactionRequest | ExecuteActorStateTransactionRequest[]>,
  ): AsyncIterable<Uint8Array> {
    for await (const pkt of source) {
      if (globalThis.Array.isArray(pkt)) {
        for (const p of (pkt as any)) {
          yield* [ExecuteActorStateTransactionRequest.encode(p).finish()];
        }
      } else {
        yield* [ExecuteActorStateTransactionRequest.encode(pkt as any).finish()];
      }
    }
  },

  // decodeTransform decodes a source of encoded messages.
  // Transform<Uint8Array, ExecuteActorStateTransactionRequest>
  async *decodeTransform(
    source: AsyncIterable<Uint8Array | Uint8Array[]> | Iterable<Uint8Array | Uint8Array[]>,
  ): AsyncIterable<ExecuteActorStateTransactionRequest> {
    for await (const pkt of source) {
      if (globalThis.Array.isArray(pkt)) {
        for (const p of (pkt as any)) {
          yield* [ExecuteActorStateTransactionRequest.decode(p)];
        }
      } else {
        yield* [ExecuteActorStateTransactionRequest.decode(pkt as any)];
      }
    }
  },

  fromJSON(object: any): ExecuteActorStateTransactionRequest {
    return {
      actorType: isSet(object.actorType) ? globalThis.String(object.actorType) : "",
      actorId: isSet(object.actorId) ? globalThis.String(object.actorId) : "",
      operations: globalThis.Array.isArray(object?.operations)
        ? object.operations.map((e: any) => TransactionalActorStateOperation.fromJSON(e))
        : [],
    };
  },

  toJSON(message: ExecuteActorStateTransactionRequest): unknown {
    const obj: any = {};
    if (message.actorType !== "") {
      obj.actorType = message.actorType;
    }
    if (message.actorId !== "") {
      obj.actorId = message.actorId;
    }
    if (message.operations?.length) {
      obj.operations = message.operations.map((e) => TransactionalActorStateOperation.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ExecuteActorStateTransactionRequest>, I>>(
    base?: I,
  ): ExecuteActorStateTransactionRequest {
    return ExecuteActorStateTransactionRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ExecuteActorStateTransactionRequest>, I>>(
    object: I,
  ): ExecuteActorStateTransactionRequest {
    const message = createBaseExecuteActorStateTransactionRequest();
    message.actorType = object.actorType ?? "";
    message.actorId = object.actorId ?? "";
    message.operations = object.operations?.map((e) => TransactionalActorStateOperation.fromPartial(e)) || [];
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends Long ? string | number | Long : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };